sturtup project

+++ src/Libraries/DoubleCode.Domain/Entity/Permissions/RolePermission.cs
@@ -9,7 +9,8 @@ public class RolePermission : BaseEntity
     public int RoleId { get; set; }
     public string PermissionTitle { get; set; }
     public string PermissionName { get; set; }
-    public int? ParentId { get; set; }
+    public long
+        ? ParentId { get; set; }
 
     #region Relations
